% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeCompositionAnalysis.R
\name{CellTypeCompositionAnalysis}
\alias{CellTypeCompositionAnalysis}
\title{Cell type composition analysis}
\usage{
CellTypeCompositionAnalysis(obs_tbl, colSample, colCelltype, colVarCats,
  colVarNums, extra_term = NULL, save = NULL)
}
\arguments{
\item{obs_tbl}{Cell level annotation table, such as AnnData.obs (data.frame-like object)}

\item{colSample}{Column in "obs_tbl" specifying sample ID (str)}

\item{colCelltype}{Column in "obs_tbl" specifying cell type annotation (str)}

\item{colVarCats}{Columns in "obs_tbl" specifying categorical variables to control for (vector of str)}

\item{colVarNums}{Columns in "obs_tbl" specifying numerical variables to control for (vector of str)}

\item{extra_term}{Extra term(s) to be added to the model (str or vector of str)}

\item{save}{If specified, used as file name prefix to save returned values (str)}
}
\value{
list containing two tables, 1) estimated random effects, 2) estimated explained variance
(in the form of standard deviation)
}
\description{
This function performs cell type composition analysis across samples grouped by certain metadata
variable while controling for other variables, by fitting a Poisson Generalised Linear Mixed
Model
}
