% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeCompositionAnalysis.R
\name{.getCondValLtsr}
\alias{.getCondValLtsr}
\title{Get mean estimate of random effects and LTSR values for all specified variables}
\usage{
.getCondValLtsr(ranef_tbl, vars = NULL, celltypes = NULL)
}
\arguments{
\item{ranef_tbl}{Table of random effect estimates (mean and sd) extracted from fitted model
(data.frame-like object)}

\item{vars}{A list specifying required categorical sample metadata variables and their order,
e.g. list(size=c('small', 'medium', 'big'), price=c('cheap', 'expensive'))}

\item{celltypes}{If specified, only keep those cell types (vector of str)}
}
\value{
tibble containing mean random effects and respective LTSR
}
\description{
This function extracts random effect estimates (mean and sd) of specified
variables, calculates LTSR using a z-score test and prepare categories in
specified orders for plotting
}
