% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeCompositionAnalysis.R
\name{.make_count_matrix}
\alias{.make_count_matrix}
\title{Make cell type by sample count matrix}
\usage{
.make_count_matrix(obs_tbl, colSample, colCelltype)
}
\arguments{
\item{obs_tbl}{Cell level annotation table, such as AnnData.obs (data.frame-like object)}

\item{colSample}{Column in "obs_tbl" specifying sample ID (str)}

\item{colCelltype}{Column in "obs_tbl" specifying cell type annotation (str)}
}
\value{
matrix containing cell type counts of dimension nSample x nCelltype
}
\description{
This function makes cell type by sample count matrix from a cell level annotation table
}
