% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeCompositionAnalysis.R
\name{.make_formula}
\alias{.make_formula}
\title{Make model formula}
\usage{
.make_formula(colSample, colVarCats, colVarNums, extra_term = NULL)
}
\arguments{
\item{colSample}{Column in "obs_tbl" specifying sample ID (str)}

\item{colVarCats}{Columns in "obs_tbl" specifying categorical variables to control for (vector of str)}

\item{colVarNums}{Columns in "obs_tbl" specifying numerical variables to control for (vector of str)}

\item{extra_term}{Extra term(s) to be added to the model (str or vector of str)}
}
\value{
formula specifying a GLMM model
}
\description{
This function makes a formula for a GLMM with random effects for each
specified categorical variables and interactions between each variable and
cell type
}
