% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeCompositionAnalysis.R
\name{.make_input_for_glmer}
\alias{.make_input_for_glmer}
\title{Make input data for model fitting}
\usage{
.make_input_for_glmer(Y, metadata_tbl)
}
\arguments{
\item{Y}{Cell type by sample count matrix (matrix of dimension nSample x nCelltype)}

\item{metadata_tbl}{Sample level metadata table (data.frame-like object)}
}
\value{
tibble containing counts and metadata for each cell type in each sample
}
\description{
This function makes input data for model fitting by assembling sample level
metadata and cell type by sample count matrix
}
