% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeCompositionAnalysis.R
\name{.make_sample_metadata}
\alias{.make_sample_metadata}
\title{Make sample level metadata table}
\usage{
.make_sample_metadata(obs_tbl, colSample, colVarCats = c(),
  colVarNums = c())
}
\arguments{
\item{obs_tbl}{Cell level annotation table, such as AnnData.obs (data.frame-like object)}

\item{colSample}{Column in "obs_tbl" specifying sample ID (str)}

\item{colVarCats}{Columns in "obs_tbl" specifying categorical variables to control for (vector of str)}

\item{colVarNums}{Columns in "obs_tbl" specifying numerical variables to control for (vector of str)}
}
\value{
tibble containing metadata for each sample, one per row
}
\description{
This function makes sample level metadata table from a cell level annotation table
}
