% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeCompositionAnalysis.R
\name{plot_sdse}
\alias{plot_sdse}
\title{Plot variable explained variance}
\usage{
plot_sdse(sdse_tbl, colSample, ci = 0.95)
}
\arguments{
\item{sdse_tbl}{Table containing square-root of estimated explained variance and standard errors
(data.frame-like object)}

\item{colSample}{Column in "obs_tbl" specifying sample ID (str)}

\item{ci}{Confidence interval level (numeric between 0 and 1, default 0.95)}
}
\value{
ggplot object of a forest plot
}
\description{
This function plots (square root of) variance explained by variables as a forest plot
}
